requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
google-generativeai==0.7.2
python-dotenv==1.0.0
aiofiles==23.2.0
pydantic==2.5.0
requests==2.31.0